toasty/tests/__init__.py: centralize some test infrastructure here
