+++ toasty/tests/test_multi_tan.py
@@ -1,5 +1,5 @@
 # -*- mode: python; coding: utf-8 -*-
-# Copyright 2019 the AAS WorldWide Telescope project
+# Copyright 2019-2020 the AAS WorldWide Telescope project
 # Licensed under the MIT License.
 
 from __future__ import absolute_import, division, print_function
@@ -10,54 +10,9 @@ import os.path
 import pytest
 import sys
 
+from . import assert_xml_elements_equal, test_path
 from .. import cli
 from .. import multi_tan
-
-
-TESTS_DIR = os.path.dirname(os.path.abspath(__file__))
-
-def make_path(*pieces):
-    return os.path.join(TESTS_DIR, *pieces)
-
-
-def check_xml_elements_equal(observed, expected):
-    """See if two XML elements are equal through recursive comparison. We do *not*
-    check the "tail" text item, and we strip whitespace in "text".
-
-    Derived om `StackExchange <https://stackoverflow.com/a/24349916/3760486>`_.
-
-    """
-    if observed.tag != expected.tag:
-        return 'expected tag {}, observed tag {}'.format(expected.tag, observed.tag)
-
-    if observed.text is None:
-        otext = ''
-    else:
-        otext = observed.text.strip()
-
-    if expected.text is None:
-        etext = ''
-    else:
-        etext = expected.text.strip()
-
-    if otext != etext:
-        return 'expected text {!r} in tag {}, observed {!r}'.format(etext, expected.tag, otext)
-    if observed.attrib != expected.attrib:
-        return 'expected attrs {!r} in tag {}, observed {!r}'.format(expected.attrib, expected.tag, observed.attrib)
-    if len(observed) != len(expected):
-        return 'expected {} children of in tag {}, observed {}'.format(len(expected), expected.tag, len(observed))
-
-    for c1, c2 in zip(observed, expected):
-        reason = check_xml_elements_equal(c1, c2)
-        if reason is not None:
-            return reason
-
-    return None
-
-def assert_xml_elements_equal(observed, expected):
-    reason = check_xml_elements_equal(observed, expected)
-    if reason is not None:
-        raise Exception('unequal XML elements: {}'.format(reason))
 
 
 class TestMultiTan(object):
@@ -111,7 +66,7 @@ class TestMultiTan(object):
         return os.path.join(self.work_dir, *pieces)
 
     def test_basic(self):
-        ds = multi_tan.MultiTanDataSource([make_path('wcs512.fits.gz')])
+        ds = multi_tan.MultiTanDataSource([test_path('wcs512.fits.gz')])
         ds.compute_global_pixelization()
 
         from xml.etree import ElementTree as etree
@@ -173,7 +128,7 @@ class TestMultiTan(object):
             'multi-tan-make-data-tiles',
             '--hdu-index', '0',
             '--outdir', self.work_path('basic_cli'),
-            make_path('wcs512.fits.gz')
+            test_path('wcs512.fits.gz')
         ]
         cli.entrypoint(args)
 
@@ -199,7 +154,7 @@ class TestMultiTan(object):
                 '--credits-text', 'CT',
                 '--credits-url', 'CU',
                 '--thumbnail-url', 'TU',
-                make_path('wcs512.fits.gz')
+                test_path('wcs512.fits.gz')
             ]
             cli.entrypoint(args)
         finally:
